Made base Item endpoints and the required code behind it to get it to work and meade seed data for storages/racks/shelves

+++ backend/SwaggerRestApi/SwaggerRestApi/Models/SpecificItem.cs
@@ -6,7 +6,7 @@
 
         public string? Description { get; set; }
 
-        public int Barcode { get; set; }
+        public Int64 Barcode { get; set; }
 
         public int? BorrowedTo { get; set; }
